// front/src/api/generated/preinversion/api.ts
export interface ErrorDetalle {
    'campo': string;
    'mensaje': string;
}
// ...
export interface ModelError {
    'codigo': string;
    'mensaje': string;
    'timestamp': string;
    'detalles'?: Array<ErrorDetalle>;
}